+++ pkg/handler/healthz.go
@@ -1,0 +1,17 @@
+// Copyright (c) 2026 Benjamin Borbe All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+package handler
+
+import (
+	"net/http"
+)
+
+// NewHealthzHandler returns an http.Handler that always responds 200 ok (liveness probe).
+func NewHealthzHandler() http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.WriteHeader(http.StatusOK)
+		_, _ = w.Write([]byte("ok"))
+	})
+}